test: split executor snapshot and role-param coverage
Move multi-role snapshot/effectiveConfig and ParamsForRoleVendor cases
out of executor_test.go into focused test files so no single agent test
file grows over the AGENTS.md 300-line design-smell threshold.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
