+++ internal/agent/executor_test.go
@@ -17,381 +17,6 @@ import (
 	"github.com/nexu-io/looper/internal/storage"
 )
 
-func TestEffectiveConfigUsesSnapshotOverrides(t *testing.T) {
-	t.Parallel()
-
-	configModel := "config-model"
-	snapshotModel := "snapshot-model"
-	executor := New(ExecutorOptions{Config: ExecutorConfig{
-		Vendor: config.AgentVendorClaudeCode,
-		Model:  &configModel,
-		Params: map[string]any{"args": []any{"--print"}},
-		Env:    map[string]string{"KEEP": "1"},
-	}})
-
-	// No snapshot: keep executor config.
-	got := executor.effectiveConfig(RunInput{})
-	if got.Vendor != config.AgentVendorClaudeCode || got.Model == nil || *got.Model != configModel {
-		t.Fatalf("effectiveConfig(no snapshot) = %#v", got)
-	}
-
-	// Snapshot overrides vendor/model for spawn only; env stays from config.
-	got = executor.effectiveConfig(RunInput{
-		UseSnapshot:    true,
-		SnapshotVendor: string(config.AgentVendorCodex),
-		SnapshotModel:  &snapshotModel,
-	})
-	if got.Vendor != config.AgentVendorCodex {
-		t.Fatalf("Vendor = %q, want codex", got.Vendor)
-	}
-	if got.Model == nil || *got.Model != snapshotModel {
-		t.Fatalf("Model = %v, want %q", got.Model, snapshotModel)
-	}
-	if got.Env["KEEP"] != "1" {
-		t.Fatalf("Env not preserved: %#v", got.Env)
-	}
-
-	// Snapshot with nil model clears model (no model flag).
-	got = executor.effectiveConfig(RunInput{
-		UseSnapshot:    true,
-		SnapshotVendor: string(config.AgentVendorOpenCode),
-		SnapshotModel:  nil,
-	})
-	if got.Vendor != config.AgentVendorOpenCode || got.Model != nil {
-		t.Fatalf("effectiveConfig(nil model) = %#v", got)
-	}
-
-	// ResolveSpawn path uses override vendor/model.
-	command, args := ResolveSpawn(got, "/tmp/wt", "hello")
-	if command != "opencode" {
-		t.Fatalf("command = %q, want opencode", command)
-	}
-	joined := strings.Join(args, " ")
-	if strings.Contains(joined, "--model") {
-		t.Fatalf("args = %q, want no --model when snapshot model is nil", joined)
-	}
-}
-
-func TestParamsForRoleVendorStripsCrossVendorIdentity(t *testing.T) {
-	t.Parallel()
-
-	global := config.AgentVendorCodex
-	roleModel := "role-model"
-	params := map[string]any{
-		"command": "/opt/codex-wrapper",
-		"args":    []any{"exec", "--model", "params-model", "--sandbox", "workspace-write"},
-		"keep":    "yes",
-	}
-
-	// Same vendor + resolved model: keep command wrapper and non-model args;
-	// strip model flags so role/profile model can win. Clone — do not mutate
-	// the global params map.
-	same := ParamsForRoleVendor(params, &global, config.AgentVendorCodex, &roleModel)
-	if same == nil || same["command"] != "/opt/codex-wrapper" {
-		t.Fatalf("same-vendor params = %#v, want original command", same)
-	}
-	sameJoined := strings.Join(stringArgs(same["args"]), " ")
-	if strings.Contains(sameJoined, "params-model") || strings.Contains(sameJoined, "--model") {
-		t.Fatalf("same-vendor args = %q, want model flags stripped when role model set", sameJoined)
-	}
-	if !strings.Contains(sameJoined, "exec") || !strings.Contains(sameJoined, "--sandbox") {
-		t.Fatalf("same-vendor args = %q, want non-model args preserved", sameJoined)
-	}
-	same["probe"] = true
-	if _, ok := params["probe"]; ok {
-		t.Fatal("same-vendor must clone params; mutated shared map")
-	}
-	// Original map must retain model flags.
-	origJoined := strings.Join(stringArgs(params["args"]), " ")
-	if !strings.Contains(origJoined, "params-model") {
-		t.Fatal("ParamsForRoleVendor mutated source params.args model flags")
-	}
-
-	// Same vendor + no resolved model: preserve params.args model flags.
-	keepModel := ParamsForRoleVendor(params, &global, config.AgentVendorCodex, nil)
-	if keepModel == nil || keepModel["command"] != "/opt/codex-wrapper" {
-		t.Fatalf("no-model same-vendor params = %#v, want original command", keepModel)
-	}
-	keepJoined := strings.Join(stringArgs(keepModel["args"]), " ")
-	if !strings.Contains(keepJoined, "params-model") || !strings.Contains(keepJoined, "--model") {
-		t.Fatalf("no-model same-vendor args = %q, want params model preserved", keepJoined)
-	}
-	keepModel["probe2"] = true
-	if _, ok := params["probe2"]; ok {
-		t.Fatal("no-model same-vendor must clone params; mutated shared map")
-	}
-
-	// Empty resolved model string is treated as unset — keep params model.
-	emptyModel := ""
-	emptyKeep := ParamsForRoleVendor(params, &global, config.AgentVendorCodex, &emptyModel)
-	emptyJoined := strings.Join(stringArgs(emptyKeep["args"]), " ")
-	if !strings.Contains(emptyJoined, "params-model") {
-		t.Fatalf("empty-model same-vendor args = %q, want params model preserved", emptyJoined)
-	}
-
-	// Diverged role vendor: drop command + all args (vendor-shaped); keep other keys.
-	diverged := ParamsForRoleVendor(params, &global, config.AgentVendorClaudeCode, &roleModel)
-	if _, ok := diverged["command"]; ok {
-		t.Fatalf("diverged role still has command: %#v", diverged)
-	}
-	if _, ok := diverged["args"]; ok {
-		t.Fatalf("diverged role still has args: %#v", diverged)
-	}
-	if diverged["keep"] != "yes" {
-		t.Fatalf("diverged role dropped non-identity param: %#v", diverged)
-	}
-	// Original map must not be mutated.
-	if params["command"] != "/opt/codex-wrapper" {
-		t.Fatal("ParamsForRoleVendor mutated source params.command")
-	}
-
-	// No global vendor: params have no home vendor; strip vendor-owned for any role.
-	noGlobal := ParamsForRoleVendor(params, nil, config.AgentVendorClaudeCode, nil)
-	if _, ok := noGlobal["command"]; ok {
-		t.Fatalf("nil global vendor still has command: %#v", noGlobal)
-	}
-	if _, ok := noGlobal["args"]; ok {
-		t.Fatalf("nil global vendor still has args: %#v", noGlobal)
-	}
-
-	// Nil params stay nil.
-	if ParamsForRoleVendor(nil, &global, config.AgentVendorClaudeCode, nil) != nil {
-		t.Fatal("nil params should stay nil")
-	}
-
-	// Role with stripped params must resolve to the role vendor binary, not the wrapper.
-	cfg := ExecutorConfig{
-		Vendor: config.AgentVendorClaudeCode,
-		Params: ParamsForRoleVendor(params, &global, config.AgentVendorClaudeCode, nil),
-	}
-	if cmd := resolveCommand(cfg); cmd != "claude" {
-		t.Fatalf("resolveCommand(diverged role) = %q, want claude", cmd)
-	}
-}
-
-func TestParamsForRoleVendorSameVendorRoleModelWinsOverParamsModelFlag(t *testing.T) {
-	t.Parallel()
-
-	global := config.AgentVendorCodex
-	roleModel := "role-model"
-	params := map[string]any{
-		"args": []any{"exec", "--model", "params-model", "--sandbox", "workspace-write"},
-	}
-	cfg := ExecutorConfig{
-		Vendor: config.AgentVendorCodex,
-		Model:  &roleModel,
-		Params: ParamsForRoleVendor(params, &global, config.AgentVendorCodex, &roleModel),
-	}
-	_, args := ResolveSpawn(cfg, "/tmp/wt", "hello")
-	joined := strings.Join(args, " ")
-	if strings.Contains(joined, "params-model") {
-		t.Fatalf("args = %q, want params model flag stripped so role model wins", joined)
-	}
-	if !strings.Contains(joined, "role-model") {
-		t.Fatalf("args = %q, want role model", joined)
-	}
-	if !strings.Contains(joined, "--sandbox") {
-		t.Fatalf("args = %q, want same-vendor non-model args preserved", joined)
-	}
-}
-
-func TestParamsForRoleVendorSameVendorPreservesParamsModelWhenNoResolvedModel(t *testing.T) {
-	t.Parallel()
-
-	global := config.AgentVendorCodex
-	params := map[string]any{
-		"args": []any{"exec", "--model", "params-model", "--sandbox", "workspace-write"},
-	}
-	cfg := ExecutorConfig{
-		Vendor: config.AgentVendorCodex,
-		Model:  nil,
-		Params: ParamsForRoleVendor(params, &global, config.AgentVendorCodex, nil),
-	}
-	_, args := ResolveSpawn(cfg, "/tmp/wt", "hello")
-	joined := strings.Join(args, " ")
-	if !strings.Contains(joined, "params-model") {
-		t.Fatalf("args = %q, want params model preserved when no resolved model", joined)
-	}
-	if !strings.Contains(joined, "--sandbox") {
-		t.Fatalf("args = %q, want same-vendor non-model args preserved", joined)
-	}
-}
-
-func TestEffectiveConfigStripsIdentityParamsUnderSnapshot(t *testing.T) {
-	t.Parallel()
-
-	configModel := "config-model"
-	snapshotModel := "frozen-model"
-	executor := New(ExecutorOptions{Config: ExecutorConfig{
-		Vendor: config.AgentVendorClaudeCode,
-		Model:  &configModel,
-		Params: map[string]any{
-			"command": "custom-agent-bin",
-			"args":    []any{"--model", "params-model", "--print", "-m", "also-params", "--other"},
-		},
-	}})
-
-	// Without snapshot: command/args params still apply.
-	got := executor.effectiveConfig(RunInput{})
-	if cmd := resolveCommand(got); cmd != "custom-agent-bin" {
-		t.Fatalf("resolveCommand(no snapshot) = %q, want custom-agent-bin", cmd)
-	}
-	command, args := ResolveSpawn(got, "/tmp/wt", "hello")
-	if command != "custom-agent-bin" {
-		t.Fatalf("command(no snapshot) = %q, want custom-agent-bin", command)
-	}
-	joined := strings.Join(args, " ")
-	if !strings.Contains(joined, "params-model") {
-		t.Fatalf("args(no snapshot) = %q, want params model flag preserved", joined)
-	}
-
-	// Snapshot vendor differs from executor: strip command + all vendor-owned args.
-	got = executor.effectiveConfig(RunInput{
-		UseSnapshot:    true,
-		SnapshotVendor: string(config.AgentVendorCodex),
-		SnapshotModel:  &snapshotModel,
-	})
-	if _, ok := got.Params["command"]; ok {
-		t.Fatalf("params still has command when snapshot vendor diverges: %#v", got.Params)
-	}
-	if _, ok := got.Params["args"]; ok {
-		t.Fatalf("params still has args when snapshot vendor diverges: %#v", got.Params)
-	}
-	if cmd := resolveCommand(got); cmd != "codex" {
-		t.Fatalf("resolveCommand(diverged snapshot) = %q, want codex", cmd)
-	}
-	command, args = ResolveSpawn(got, "/tmp/wt", "hello")
-	if command != "codex" {
-		t.Fatalf("command(diverged snapshot) = %q, want codex", command)
-	}
-	joined = strings.Join(args, " ")
-	if strings.Contains(joined, "params-model") || strings.Contains(joined, "also-params") || strings.Contains(joined, "--other") {
-		t.Fatalf("args(snapshot) = %q, want clean vendor defaults (no foreign params.args)", joined)
-	}
-	if !strings.Contains(joined, "frozen-model") {
-		t.Fatalf("args(snapshot) = %q, want frozen model", joined)
-	}
-
-	// Same vendor as executor: keep params.command wrapper, still strip model flags.
-	got = executor.effectiveConfig(RunInput{
-		UseSnapshot:    true,
-		SnapshotVendor: string(config.AgentVendorClaudeCode),
-		SnapshotModel:  &snapshotModel,
-	})
-	if cmd := resolveCommand(got); cmd != "custom-agent-bin" {
-		t.Fatalf("resolveCommand(same-vendor snapshot) = %q, want custom-agent-bin wrapper", cmd)
-	}
-	command, args = ResolveSpawn(got, "/tmp/wt", "hello")
-	if command != "custom-agent-bin" {
-		t.Fatalf("command(same-vendor snapshot) = %q, want custom-agent-bin", command)
-	}
-	joined = strings.Join(args, " ")
-	if strings.Contains(joined, "params-model") || strings.Contains(joined, "also-params") {
-		t.Fatalf("args(same-vendor snapshot) = %q, want model flags stripped", joined)
-	}
-	if !strings.Contains(joined, "frozen-model") {
-		t.Fatalf("args(same-vendor snapshot) = %q, want frozen model", joined)
-	}
-
-	// Original executor params must not be mutated.
-	if _, ok := executor.config.Params["command"]; !ok {
-		t.Fatal("executor config params.command was mutated")
-	}
-}
-
-func TestEffectiveConfigKeepsGlobalParamsWhenSnapshotMatchesOwner(t *testing.T) {
-	t.Parallel()
-
-	// Global Codex owns params.command; the live role handler is Claude after a
-	// hot switch. Sticky retry still restores Codex from agent_snapshot_json.
-	global := config.AgentVendorCodex
-	roleModel := "role-claude-model"
-	snapshotModel := "frozen-codex-model"
-	params := map[string]any{
-		"command": "/opt/codex-wrapper",
-		"args":    []any{"exec", "--model", "params-model", "--sandbox", "workspace-write"},
-		"keep":    "yes",
-	}
-	executor := New(ExecutorOptions{
-		Config: ExecutorConfig{
-			Vendor: config.AgentVendorClaudeCode,
-			Model:  &roleModel,
-			Params: params,
-		},
-		ParamsOwnerVendor: &global,
-	})
-
-	// Live role claim (no snapshot): strip Codex wrappers so Claude launches.
-	got := executor.effectiveConfig(RunInput{})
-	if _, ok := got.Params["command"]; ok {
-		t.Fatalf("live diverged role still has command: %#v", got.Params)
-	}
-	if _, ok := got.Params["args"]; ok {
-		t.Fatalf("live diverged role still has args: %#v", got.Params)
-	}
-	if got.Params["keep"] != "yes" {
-		t.Fatalf("live diverged role dropped non-identity param: %#v", got.Params)
-	}
-	if cmd := resolveCommand(got); cmd != "claude" {
-		t.Fatalf("resolveCommand(live role) = %q, want claude", cmd)
-	}
-
-	// Sticky snapshot restores Codex (params owner): keep wrapper + non-model args.
-	got = executor.effectiveConfig(RunInput{
-		UseSnapshot:    true,
-		SnapshotVendor: string(config.AgentVendorCodex),
-		SnapshotModel:  &snapshotModel,
-	})
-	if got.Params["command"] != "/opt/codex-wrapper" {
-		t.Fatalf("snapshot matching owner lost command: %#v", got.Params)
-	}
-	if cmd := resolveCommand(got); cmd != "/opt/codex-wrapper" {
-		t.Fatalf("resolveCommand(snapshot owner) = %q, want /opt/codex-wrapper", cmd)
-	}
-	command, args := ResolveSpawn(got, "/tmp/wt", "hello")
-	if command != "/opt/codex-wrapper" {
-		t.Fatalf("command(snapshot owner) = %q, want /opt/codex-wrapper", command)
-	}
-	joined := strings.Join(args, " ")
-	if strings.Contains(joined, "params-model") || strings.Contains(joined, "--model params-model") {
-		t.Fatalf("args(snapshot owner) = %q, want params model flags stripped", joined)
-	}
-	if !strings.Contains(joined, "frozen-codex-model") {
-		t.Fatalf("args(snapshot owner) = %q, want frozen snapshot model", joined)
-	}
-	if !strings.Contains(joined, "exec") || !strings.Contains(joined, "--sandbox") {
-		t.Fatalf("args(snapshot owner) = %q, want non-model args preserved", joined)
-	}
-
-	// Snapshot to a third vendor still strips owner wrappers.
-	got = executor.effectiveConfig(RunInput{
-		UseSnapshot:    true,
-		SnapshotVendor: string(config.AgentVendorOpenCode),
-		SnapshotModel:  &snapshotModel,
-	})
-	if _, ok := got.Params["command"]; ok {
-		t.Fatalf("third-vendor snapshot still has command: %#v", got.Params)
-	}
-	if cmd := resolveCommand(got); cmd != "opencode" {
-		t.Fatalf("resolveCommand(third-vendor snapshot) = %q, want opencode", cmd)
-	}
-
-	// Shared global params map must not be mutated.
-	if params["command"] != "/opt/codex-wrapper" {
-		t.Fatal("effectiveConfig mutated shared global params.command")
-	}
-}
-
-func TestStripModelFlags(t *testing.T) {
-	t.Parallel()
-
-	got := stripModelFlags([]string{"--model", "x", "-m", "y", "--model=z", "-m=w", "-mMODEL", "--keep", "v"})
-	want := []string{"--keep", "v"}
-	if strings.Join(got, " ") != strings.Join(want, " ") {
-		t.Fatalf("stripModelFlags = %v, want %v", got, want)
-	}
-}
-
 func TestResolveSpawnVendorParity(t *testing.T) {
 	t.Parallel()
